Merged PR 10974: Add docker compose, update package-lock.json
Also removed a dependency in projects-by-type.component.ts that was causing errors when serving mango

**Note:**
When first pulling down the new repo, had issues serving the app. Needed to delete package-lock.json and run npm i to fix the issue.

Adding the updated package-lock.json file to the PR so others dont have this issue.