Merged PR 20978: 155143: BUG: CA - Service account - Generate API key button broken in STAGE
The **issue** is due to a new feature that was added for MangoSPA on a **shared component** that the CA service account page also uses (e.g. the shared Modal component crem-modal).

Forced to bring in DataIdBreadcrumbProviderService and MangoAppFacade into CA's SPA application since its using the Modal module.
- Ideally, whoever worked on this should take a look to maybe remove this from the shared modal?

Related work items: #155143